KS-11-1 fix
